# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: redeluge JSON-RPC API
  version: "2.2.1"
  summary: The HTTP API of the redeluge Web UI server.
  description:
    |-
      Everything a client does goes through one endpoint, `POST /json`, carrying
      JSON-RPC version 1. This is Deluge's own API, unchanged: redeluge reproduces
      it method for method so that clients written against the Python server keep
      working. The version above is the API's, not the fork's.

      A call is `{"method": ..., "params": [...], "id": ...}` and a reply is always
      HTTP 200 with `{"result": ..., "error": ..., "id": ...}`. A failure is
      reported in `error`, never in the status code, because that is what the
      shipped front end expects.

      Three methods answer before a session exists: `auth.login`,
      `auth.check_session` and `system.listMethods`. Everything else needs the
      session cookie that `auth.login` sets.

      Methods in the `core.*` and `daemon.*` namespaces are forwarded to the daemon
      over DelugeRPC; the rest are answered by the Web UI server itself.
  license:
    name: GPL-3.0-or-later
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: http://localhost:8112
    description: The Web UI, as the container publishes it.
...
security:
  - sessionCookie: []
paths:
  /json:
    post:
      summary: Call one method.
      description:
        |-
          The body names the method and its positional parameters. The response is
          HTTP 200 whatever happens; read `error` to find out whether the call worked.

          `auth.login` is the only method that mints a session: it sets the
          `_session_id` cookie, which every later call must carry.
      operationId: call
      tags: [auth, core, daemon, system, web, webutils]
      security:
        - sessionCookie: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Request"
      responses:
        "200":
          description: The call was dispatched. `error` says whether it worked.
          headers:
            Set-Cookie:
              description: Sent by `auth.login`, and refreshed on every call.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "400":
          description: The body was not a JSON-RPC call.
  /upload:
    post:
      summary: Stage one or more torrent files.
      description: |-
        The add-by-file dialog posts a multipart form here. The answer carries
        the paths the files were staged at, which then go to
        `web.get_torrent_info` and `web.add_torrents`.

        Always HTTP 200 with a `success` field: ExtJS's form submit treats any
        other status as a transport failure and shows its own message instead
        of this one. A file that does not parse as a torrent is refused, and
        one file may not exceed 10 MB.
      operationId: upload
      tags: [web]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: Whether the files were staged, and where.
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
                  files:
                    type: array
                    description: Paths to pass to `web.get_torrent_info`.
                    items:
                      type: string
                  error:
                    type: string
...
  /render/{name}:
    get:
      summary: One rendered template fragment.
      description:
        |-
          The front end fetches a few fragments of markup rather than building them in
          JavaScript. Served from the templates compiled into the binary.
      operationId: render
      tags: [web]
      security:
        - {}
      parameters:
        - name: name
          in: path
          required: true
          description: The fragment's file name, such as `tab_status.html`.
          schema:
            type: string
      responses:
        "200":
          description: The fragment.
          content:
            text/html:
              schema:
                type: string
        "404":
          description: No such fragment.
  /flag/{code}:
    get:
      summary: The flag for a peer's country.
      description:
        |-
          The peers tab draws one of these per row. The code is ISO 3166-1 alpha-2 and
          is validated: anything else is a 404 rather than a path. Countries are only
          filled in when the daemon has a GeoIP database configured, so on a default
          install this is never asked for.
      operationId: flag
      tags: [web]
      security:
        - {}
      parameters:
        - name: code
          in: path
          required: true
          description: A two-letter country code, such as `fr`.
          schema:
            type: string
            pattern: "^[A-Za-z]{2}$"
      responses:
        "200":
          description: The flag.
          content:
            image/png:
              schema:
                type: string
                format: binary
        "404":
          description: No flag for that code.
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: _session_id
      description:
        |-
          Set by `auth.login`. The value is a session id with a four-digit checksum
          appended, which is the shape the Python server used; the checksum is not a
          security control, the id behind it is.
  schemas:
...
    Error:
      type: object
      description: Present on `error` when a call failed, and `null` otherwise.
      required: [message, code]
      properties:
        message:
          type: string
          description: What went wrong, in words.
        code:
          type: integer
          description:
            |-
              1 for a call that needed a session and did not have one, 2 for a method this
              server does not know, 3 for a failure inside the method, 4 for an error the
              daemon reported.
    Response:
      type: object
      description: Always HTTP 200. Exactly one of `result` and `error` is set.
      required: [result, error, id]
      properties:
        result:
          description: Whatever the method returns. `null` when it failed.
        error:
          oneOf:
            - $ref: "#/components/schemas/Error"
            - type: "null"
        id:
          $ref: "#/components/schemas/RequestId"
    Request:
      description: One call. The `method` field picks which shape applies.
      discriminator:
        propertyName: method
      oneOf:
        - $ref: "#/components/schemas/AuthChangePassword"
        - $ref: "#/components/schemas/AuthCheckSession"
        - $ref: "#/components/schemas/AuthDeleteSession"
        - $ref: "#/components/schemas/AuthLogin"
        - $ref: "#/components/schemas/CoreAddTorrentFile"
        - $ref: "#/components/schemas/CoreAddTorrentFileAsync"
        - $ref: "#/components/schemas/CoreAddTorrentFiles"
        - $ref: "#/components/schemas/CoreAddTorrentMagnet"
        - $ref: "#/components/schemas/CoreAddTorrentUrl"
        - $ref: "#/components/schemas/CoreConnectPeer"
        - $ref: "#/components/schemas/CoreCreateAccount"
        - $ref: "#/components/schemas/CoreCreateTorrent"
        - $ref: "#/components/schemas/CoreForceReannounce"
        - $ref: "#/components/schemas/CoreForceRecheck"
        - $ref: "#/components/schemas/CoreGetAuthLevelsMappings"
        - $ref: "#/components/schemas/CoreGetCompletionPaths"
        - $ref: "#/components/schemas/CoreGetConfig"
        - $ref: "#/components/schemas/CoreGetConfigValue"
        - $ref: "#/components/schemas/CoreGetConfigValues"
        - $ref: "#/components/schemas/CoreGetExternalIp"
        - $ref: "#/components/schemas/CoreGetFilterTree"
        - $ref: "#/components/schemas/CoreGetFreeSpace"
        - $ref: "#/components/schemas/CoreGetKnownAccounts"
        - $ref: "#/components/schemas/CoreGetLibtorrentVersion"
        - $ref: "#/components/schemas/CoreGetListenPort"
        - $ref: "#/components/schemas/CoreGetMagnetUri"
        - $ref: "#/components/schemas/CoreGetPathSize"
        - $ref: "#/components/schemas/CoreGetProxy"
        - $ref: "#/components/schemas/CoreGetSessionState"
        - $ref: "#/components/schemas/CoreGetSessionStatus"
        - $ref: "#/components/schemas/CoreGetSslListenPort"
        - $ref: "#/components/schemas/CoreGetTorrentStatus"
        - $ref: "#/components/schemas/CoreGetTorrentsStatus"
        - $ref: "#/components/schemas/CoreGlob"
        - $ref: "#/components/schemas/CoreIsSessionPaused"
        - $ref: "#/components/schemas/CoreMoveStorage"
        - $ref: "#/components/schemas/CorePauseSession"
        - $ref: "#/components/schemas/CorePauseTorrent"
        - $ref: "#/components/schemas/CorePauseTorrents"
        - $ref: "#/components/schemas/CorePrefetchMagnetMetadata"
        - $ref: "#/components/schemas/CoreQueueBottom"
        - $ref: "#/components/schemas/CoreQueueDown"
        - $ref: "#/components/schemas/CoreQueueTop"
        - $ref: "#/components/schemas/CoreQueueUp"
        - $ref: "#/components/schemas/CoreRemoveAccount"
        - $ref: "#/components/schemas/CoreRemoveTorrent"
        - $ref: "#/components/schemas/CoreRemoveTorrents"
        - $ref: "#/components/schemas/CoreRenameFiles"
        - $ref: "#/components/schemas/CoreRenameFolder"
        - $ref: "#/components/schemas/CoreResumeSession"
        - $ref: "#/components/schemas/CoreResumeTorrent"
        - $ref: "#/components/schemas/CoreResumeTorrents"
        - $ref: "#/components/schemas/CoreSetConfig"
        - $ref: "#/components/schemas/CoreSetSslTorrentCert"
        - $ref: "#/components/schemas/CoreSetTorrentAutoManaged"
        - $ref: "#/components/schemas/CoreSetTorrentFilePriorities"
        - $ref: "#/components/schemas/CoreSetTorrentMaxConnections"
        - $ref: "#/components/schemas/CoreSetTorrentMaxDownloadSpeed"
        - $ref: "#/components/schemas/CoreSetTorrentMaxUploadSlots"
        - $ref: "#/components/schemas/CoreSetTorrentMaxUploadSpeed"
        - $ref: "#/components/schemas/CoreSetTorrentMoveCompleted"
        - $ref: "#/components/schemas/CoreSetTorrentMoveCompletedPath"
        - $ref: "#/components/schemas/CoreSetTorrentOptions"
        - $ref: "#/components/schemas/CoreSetTorrentPrioritizeFirstLast"
        - $ref: "#/components/schemas/CoreSetTorrentRemoveAtRatio"
        - $ref: "#/components/schemas/CoreSetTorrentStopAtRatio"
        - $ref: "#/components/schemas/CoreSetTorrentStopRatio"
        - $ref: "#/components/schemas/CoreSetTorrentTrackers"
        - $ref: "#/components/schemas/CoreTestListenPort"
        - $ref: "#/components/schemas/CoreUpdateAccount"
        - $ref: "#/components/schemas/DaemonAuthorizedCall"
        - $ref: "#/components/schemas/DaemonGetMethodList"
        - $ref: "#/components/schemas/DaemonGetVersion"
        - $ref: "#/components/schemas/DaemonShutdown"
        - $ref: "#/components/schemas/SystemListMethods"
        - $ref: "#/components/schemas/WebAddHost"
        - $ref: "#/components/schemas/WebAddTorrents"
        - $ref: "#/components/schemas/WebConnect"
        - $ref: "#/components/schemas/WebConnected"
        - $ref: "#/components/schemas/WebDeregisterEventListener"
        - $ref: "#/components/schemas/WebDisconnect"
        - $ref: "#/components/schemas/WebDownloadTorrentFromUrl"
        - $ref: "#/components/schemas/WebEditHost"
        - $ref: "#/components/schemas/WebGetConfig"
        - $ref: "#/components/schemas/WebGetEvents"
        - $ref: "#/components/schemas/WebGetHostStatus"
        - $ref: "#/components/schemas/WebGetHosts"
        - $ref: "#/components/schemas/WebGetMagnetInfo"
        - $ref: "#/components/schemas/WebGetTorrentFiles"
        - $ref: "#/components/schemas/WebGetTorrentInfo"
        - $ref: "#/components/schemas/WebGetTorrentStatus"
        - $ref: "#/components/schemas/WebRegisterEventListener"
        - $ref: "#/components/schemas/WebRemoveHost"
        - $ref: "#/components/schemas/WebSetConfig"
        - $ref: "#/components/schemas/WebSetTheme"
        - $ref: "#/components/schemas/WebStartDaemon"
        - $ref: "#/components/schemas/WebStopDaemon"
        - $ref: "#/components/schemas/WebUpdateUi"
        - $ref: "#/components/schemas/WebutilsGetLanguages"
        - $ref: "#/components/schemas/WebutilsGetThemes"
    AuthChangePassword:
      type: object
      description:
        |-
          Change the password.
          Authorisation: normal.
          Answered by the Web UI server.
          Required parameters, in order: `old_password`, `new_password`.
          Defined in the Python implementation at `deluge/ui/web/auth.py:205`.
      required: [method]
      properties:
        method:
          type: string
          const: "auth.change_password"
        id:
          $ref: "#/components/schemas/RequestId"
        params:
          type: array
          description: Positional, in the order below.
          minItems: 2
          maxItems: 2
          prefixItems:
            - description: "old_password"
            - description: "new_password"
    AuthCheckSession:
      type: object
      description:
        |-
          Check a session to see if it's still valid.
          Authorisation: none, answered before a session exists.
          Answered by the Web UI server.
          Optional, in order after those: `session_id` (default `None`).
          Defined in the Python implementation at `deluge/ui/web/auth.py:219`.
      required: [method]
      properties:
        method:
          type: string
          const: "auth.check_session"
        id:
          $ref: "#/components/schemas/RequestId"
        params:
          type: array
          description: Positional, in the order below.
          minItems: 0
          maxItems: 1
          prefixItems:
            - description: "session_id — optional, defaults to None"
    AuthDeleteSession:
      type: object
      description:
        |-
          Removes a session.
          Authorisation: normal.
          Answered by the Web UI server.
          Takes no parameters.
          Defined in the Python implementation at `deluge/ui/web/auth.py:229`.
      required: [method]
      properties:
        method:
          type: string
          const: "auth.delete_session"
        id:
          $ref: "#/components/schemas/RequestId"
        params:
          type: array
          description: Empty; this method takes no parameters.
          maxItems: 0
          default: []
    AuthLogin:
      type: object
      description:
        |-
          Test a password to see if it's valid.
          Authorisation: none, answered before a session exists.
          Answered by the Web UI server.
          Required parameters, in order: `password`.
          Defined in the Python implementation at `deluge/ui/web/auth.py:240`.
      required: [method]
      properties:
        method:
          type: string
          const: "auth.login"
        id:
          $ref: "#/components/schemas/RequestId"
        params:
          type: array
          description: Positional, in the order below.
          minItems: 1
          maxItems: 1
          prefixItems:
            - description: "password"
    CoreAddTorrentFile:
      type: object
      description:
        |-
          Adds a torrent file to the session.
          Authorisation: normal.
          Forwarded to the daemon over DelugeRPC.
          Required parameters, in order: `filename`, `filedump`, `options`.
          Defined in the Python implementation at `deluge/core/core.py:457`.
      required: [method]
      properties:
        method:
          type: string
          const: "core.add_torrent_file"
        id:
          $ref: "#/components/schemas/RequestId"
        params:
          type: array
          description: Positional, in the order below.
          minItems: 3
          maxItems: 3
          prefixItems:
            - description: "filename (str)"
            - description: "filedump (Union[str, bytes])"
            - description: "options (dict)"
    CoreAddTorrentFileAsync:
      type: object
      description:
        |-
          Adds a torrent file to the session asynchronously.
          Authorisation: normal.
          Forwarded to the daemon over DelugeRPC.
          Required parameters, in order: `filename`, `filedump`, `options`.
          Optional, in order after those: `save_state` (default `True`).
          Defined in the Python implementation at `deluge/core/core.py:403`.
      required: [method]
      properties:
        method:
          type: string
          const: "core.add_torrent_file_async"
        id:
          $ref: "#/components/schemas/RequestId"
        params:
          type: array
          description: Positional, in the order below.
          minItems: 3
          maxItems: 4
          prefixItems:
            - description: "filename (str)"
            - description: "filedump (str)"
            - description: "options (dict)"
            - description: "save_state (bool) — optional, defaults to True"
    CoreAddTorrentFiles:
      type: object
      description:
        |-
          Adds multiple torrent files to the session asynchronously.
          Authorisation: normal.
          Forwarded to the daemon over DelugeRPC.
          Required parameters, in order: `torrent_files`.
          Defined in the Python implementation at `deluge/core/core.py:484`.
      required: [method]
      properties:
        method:
          type: string
          const: "core.add_torrent_files"
        id:
          $ref: "#/components/schemas/RequestId"
        params:
          type: array
          description: Positional, in the order below.
          minItems: 1
          maxItems: 1
          prefixItems:
            - description: "torrent_files (List[Tuple[str, Union[str, bytes], dict]])"
    CoreAddTorrentMagnet:
      type: object
      description:
        |-
          Adds a torrent from a magnet link.
          Authorisation: normal.
          Forwarded to the daemon over DelugeRPC.
          Required parameters, in order: `uri`, `options`.
          Defined in the Python implementation at `deluge/core/core.py:551`.
      required: [method]
      properties:
        method:
          type: string
          const: "core.add_torrent_magnet"
        id:
          $ref: "#/components/schemas/RequestId"
        params:
          type: array
          description: Positional, in the order below.
          minItems: 2
          maxItems: 2
          prefixItems:
            - description: "uri (str)"
            - description: "options (dict)"
    CoreAddTorrentUrl:
      type: object
      description:
        |-
          Adds a torrent from a URL. Deluge will attempt to fetch the torrent
          Authorisation: normal.
          Forwarded to the daemon over DelugeRPC.
          Required parameters, in order: `url`, `options`.
          Optional, in order after those: `headers` (default `None`).
          Defined in the Python implementation at `deluge/core/core.py:515`.
      required: [method]
      properties:
        method:
          type: string
          const: "core.add_torrent_url"
        id:
          $ref: "#/components/schemas/RequestId"
        params:
          type: array
          description: Positional, in the order below.
          minItems: 2
          maxItems: 3
          prefixItems:
            - description: "url (str)"
            - description: "options (dict)"
            - description: "headers (dict) — optional, defaults to None"
    CoreConnectPeer:
      type: object
      description:
        |-
          Authorisation: normal.
          Forwarded to the daemon over DelugeRPC.
          Required parameters, in order: `torrent_id`, `ip`, `port`.
          Defined in the Python implementation at `deluge/core/core.py:674`.
      required: [method]
      properties:
        method:
          type: string
          const: "core.connect_peer"
        id:
          $ref: "#/components/schemas/RequestId"
        params:
          type: array
          description: Positional, in the order below.
          minItems: 3
          maxItems: 3
          prefixItems:
            - description: "torrent_id (str)"
            - description: "ip (str)"
            - description: "port (int)"
    CoreCreateAccount:
      type: object
      description:
        |-
          Authorisation: admin.
          Forwarded to the daemon over DelugeRPC.
          Required parameters, in order: `username`, `password`, `authlevel`.
          Defined in the Python implementation at `deluge/core/core.py:1361`.
      required: [method]
      properties:
        method:
          type: string
          const: "core.create_account"
        id:
          $ref: "#/components/schemas/RequestId"
        params:
          type: array
          description: Positional, in the order below.
          minItems: 3
          maxItems: 3
          prefixItems:
            - description: "username (str)"
            - description: "password (str)"
            - description: "authlevel (str)"
    CoreCreateTorrent:
      type: object
      description:
        |-
          Authorisation: normal.
          Forwarded to the daemon over DelugeRPC.
          Required parameters, in order: `path`, `tracker`, `piece_length`.
          Optional, in order after those: `comment` (default `None`), `target` (default `None`), `webseeds` (default `None`), `private` (default `False`), `created_by` (default `None`), `trackers` (default `None`), `add_to_session` (default `False`), `torrent_format` (default `metafile.TorrentFormat.V1`), `ca_cert` (default `None`).
          Defined in the Python implementation at `deluge/core/core.py:1053`.
      required: [method]
      properties:
        method:
          type: string
          const: "core.create_torrent"
        id:
          $ref: "#/components/schemas/RequestId"
        params:
          type: array
          description: Positional, in the order below.
          minItems: 3
          maxItems: 12
          prefixItems:
            - description: "path"
            - description: "tracker"
            - description: "piece_length"
            - description: "comment — optional, defaults to None"
            - description: "target — optional, defaults to None"
            - description: "webseeds — optional, defaults to None"
            - description: "private — optional, defaults to False"
            - description: "created_by — optional, defaults to None"
            - description: "trackers — optional, defaults to None"
            - description: "add_to_session — optional, defaults to False"
            - description: "torrent_format — optional, defaults to metafile.TorrentFormat.V1"
            - description: "ca_cert — optional, defaults to None"
    CoreForceReannounce:
      type: object
      description:
        |-
          Authorisation: normal.
          Forwarded to the daemon over DelugeRPC.
          Required parameters, in order: `torrent_ids`.
          Defined in the Python implementation at `deluge/core/core.py:651`.
      required: [method]
      properties:
        method:
          type: string
          const: "core.force_reannounce"
        id:
          $ref: "#/components/schemas/RequestId"
        params:
          type: array
          description: Positional, in the order below.
          minItems: 1
          maxItems: 1
          prefixItems:
            - description: "torrent_ids (List[str])"
    CoreForceRecheck:
      type: object
      description:
        |-
          Forces a data recheck on torrent_ids
          Authorisation: normal.
          Forwarded to the daemon over DelugeRPC.
          Required parameters, in order: `torrent_ids`.
          Defined in the Python implementation at `deluge/core/core.py:938`.
      required: [method]
      properties:
        method:
          type: string
          const: "core.force_recheck"
        id:
          $ref: "#/components/schemas/RequestId"
        params:
          type: array
          description: Positional, in the order below.
          minItems: 1
          maxItems: 1
          prefixItems:
            - description: "torrent_ids (List[str])"
    CoreGetAuthLevelsMappings:
      type: object
      description:
        |-
          Authorisation: none, answered before a session exists.
          Forwarded to the daemon over DelugeRPC.
          Takes no parameters.
          Defined in the Python implementation at `deluge/core/core.py:1357`.
      required: [method]
      properties:
        method:
          type: string
          const: "core.get_auth_levels_mappings"
        id:
          $ref: "#/components/schemas/RequestId"
        params:
          type: array
          description: Empty; this method takes no parameters.
          maxItems: 0
          default: []
    CoreGetCompletionPaths:
      type: object
      description:
        |-
          Returns the available path completions for the input value.
          Authorisation: normal.
          Forwarded to the daemon over DelugeRPC.
          Required parameters, in order: `args`.
          Defined in the Python implementation at `deluge/core/core.py:1348`.
      required: [method]
      properties:
        method:
          type: string
          const: "core.get_completion_paths"
        id:
          $ref: "#/components/schemas/RequestId"
        params:
          type: array
          description: Positional, in the order below.
          minItems: 1
          maxItems: 1
          prefixItems:
            - description: "args (Dict[str, Any])"
    CoreGetConfig:
      type: object
      description:
        |-
          Get all the preferences as a dictionary
          Authorisation: normal.
          Forwarded to the daemon over DelugeRPC.
          Takes no parameters.
          Defined in the Python implementation at `deluge/core/core.py:849`.
      required: [method]
      properties:
        method:
          type: string
          const: "core.get_config"
        id:
          $ref: "#/components/schemas/RequestId"
        params:
          type: array
          description: Empty; this method takes no parameters.
          maxItems: 0
          default: []
    CoreGetConfigValue:
      type: object
      description:
        |-
          Get the config value for key
          Authorisation: normal.
          Forwarded to the daemon over DelugeRPC.
          Required parameters, in order: `key`.
          Defined in the Python implementation at `deluge/core/core.py:854`.
      required: [method]
      properties:
        method:
          type: string
          const: "core.get_config_value"
        id:
          $ref: "#/components/schemas/RequestId"
        params:
          type: array
          description: Positional, in the order below.
          minItems: 1
          maxItems: 1
          prefixItems:
            - description: "key (str)"
...